backend/requirements.txt
fastapi>=0.96.0,<0.97.0
pydantic>=1.10.0,<1.11.0
uvicorn>=0.22.0,<0.23.0
pytest>=7.3.0,<7.4.0
flake8>=6.0.0,<6.1.0
psycopg2>=2.9.0,<3.0.0
python-dotenv>=1.0.0,<1.1.0
sqlalchemy>=2.0.0,<2.1.0
httpx>=0.24.1,<0.25.0
minio>=7.1.15,<7.2.0
python-multipart>=0.0.5,<0.0.6
passlib>=1.7.3,<1.7.4
python-jose>=3.2.0,<3.3.0